+++ ActionCompiler/AST/Types/FloatNode.cs
@@ -1,6 +1,6 @@
 ﻿namespace Action.AST
 {
-    public record FloatNode(float f) : ValueNode
+    public record FloatNode(double f) : ValueNode
     {
         public override T Accept<T>(NodeVisitor<T> visitor)
         {
